Moving specs to root of the project and renaming to 'test'
