+++ .gitignore
@@ -4,3 +4,4 @@ moip-sdk-node.iml
 .DS_Store
 .nyc_output
 dist/
+test_dist/
