+++ lib/client/api.js
@@ -9,7 +9,8 @@ const get = (opts, endpoint, id, config) => {
             'Authorization': opts.auth
         },
         method: 'GET',
-        json: true
+        json: true,
+        resolveWithFullResponse: true
     };
 
     return request(options);
@@ -24,10 +25,9 @@ const post = (opts, endpoint, payload, config) => {
         method: 'POST',
         body: payload,
         form: config && config.form,
-        json: !config || !config.form
+        json: !config || !config.form,
+        resolveWithFullResponse: true
     };
-
-    console.log(options);
 
     return request(options);
 };
@@ -39,7 +39,8 @@ const remove = (opts, endpoint) => {
             'Authorization': opts.auth
         },
         method: 'DELETE',
-        json: true
+        json: true,
+        resolveWithFullResponse: true
     };
 
     return request(options);
