+++ test/account.spec.js
@@ -1,5 +1,5 @@
 import auth from './config/auth';
-import moip from '../index';
+import moip from '../dist/index';
 import chai from 'chai';
 import generateCPF from 'gerar-cpf';
 import accountModel from './schemas/account';
@@ -16,7 +16,8 @@ describe('Account', () => {
     it('Should successfully create an account', (done) => {
         moip.init(auth).then((client) => {
             client.account.create(accountModel)
-            .then(() => {
+            .then((response) => {
+                response.statusCode.should.be.eql(201);
                 done();
             }).catch((err) => done(err.statusCode));
         });
