+++ test/config/auth.js
