+++ test/config/basicAuth.js
