+++ test/coupon.spec.js
@@ -1,8 +1,8 @@
-var auth = require('specs/config/auth.js')
+var auth = require('test/config/auth.js')
 var moip = require('index.js')(auth)
-var subscription = require('specs/schemas/subscription.js')
-var plan = require('specs/schemas/plan.js')
-var coupon = require('specs/schemas/coupon.js')
+var subscription = require('test/schemas/subscription.js')
+var plan = require('test/schemas/plan.js')
+var coupon = require('test/schemas/coupon.js')
 var shortid = require('shortid')
 var chai = require('chai')
 chai.should()
