+++ test/index.js
@@ -1,0 +1,18 @@
+/* Define testing order */
+import 'babel-polyfill';
+
+import './account.spec';
+import './customer.spec';
+import './order.spec';
+import './payment.spec';
+import './refund.spec';
+import './notification.spec';
+import './bankAccount.spec';
+import './escrow.spec';
+import './webhook.spec';
+import './connect.spec';
+
+// require('./plan.spec.js');
+// require('./subscriber.spec.js');
+// require('./subscription.spec.js');
+// require('./coupon.spec.js');
