+++ test/schemas/subscription.js
