+++ test/webhook.spec.js
@@ -1,5 +1,5 @@
 import auth from './config/auth';
-import moip from '../index';
+import moip from '../dist/index';
 import chai from 'chai';
 
 chai.should();
@@ -13,7 +13,7 @@ describe('Moip Webhooks', () => {
     it('Should successfully get all webhooks', (done) => {
         moip.init(auth).then((client) => {
             client.webhook.getAll()
-                .then((body) => {
+                .then(({body}) => {
                     body.should.have.property('webhooks');
                     webhook.resourceId = body.webhooks[0].resourceId;
                     webhook.id = body.webhooks[0].id;
@@ -25,7 +25,7 @@ describe('Moip Webhooks', () => {
     it('Should successfully get webhooks from specific resource', (done) => {
         moip.init(auth).then((client) => {
             client.webhook.getOne(webhook.resourceId)
-                .then((body) => {
+                .then(({body}) => {
                     body.should.have.property('webhooks');
                     done();
                 })
